Minor refactoring

+++ sdev/include/shaiya/include/UserHelper.h
@@ -12,10 +12,10 @@ namespace shaiya
     {
         static bool HasApplySkill(CUser* user, int skillId, int skillLv);
 
-        static bool ItemCreate(CUser* user, ItemInfo* itemInfo, int count, int& outBag, int& outSlot);
-        static bool ItemRemove(CUser* user, int bag, int slot, int count);
-        static bool ItemRemove(CUser* user, ItemId itemId, int count);
-        static bool ItemRemove(CUser* user, ItemEffect effect, int count);
+        static bool ItemCreate(CUser* user, ItemInfo* itemInfo, uint8_t count, int& outBag, int& outSlot);
+        static bool ItemRemove(CUser* user, uint8_t bag, uint8_t slot, uint8_t count);
+        static bool ItemRemove(CUser* user, ItemId itemId, uint8_t count);
+        static bool ItemRemove(CUser* user, ItemEffect effect, uint8_t count);
 
         static void SetMovePosition(CUser* user, int mapId, float x, float y, float z, int recallType, ULONG delay);
         static void SetMovePosition(CUser* user, int mapId, SVector* pos, int recallType, ULONG delay);
